LayoutTests/js/dom/unhandled-promise-rejection-handle-in-workers-expected.txt

[Worker] Test rejection handled event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/unhandled-promise-rejection-handle-in-workers.js
PASS [Worker] error.type is "unhandledrejection"
PASS [Worker] error.cancelable is true
PASS [Worker] error.promise is promise
PASS [Worker] error.reason is "ERROR"
PASS [Worker] reason is "ERROR"
PASS [Worker] handled.type is "rejectionhandled"
PASS [Worker] handled.cancelable is false
PASS [Worker] handled.promise is promise
PASS [Worker] handled.reason is "ERROR"
PASS successfullyParsed is true

TEST COMPLETE
